The Catalog
The Catalog is updated regularly with the latest data from the LIMS for all the studies at the IBRH3AU Biorepository, We're trusted by many leading research institutions, academic institutes and biomedical companies for diagnostics, health care or basic research. Through the catalog you are able to search and request for biospecimens or human tissues/organs for health care or basic research needs.